Changes to modelCrud.test, fixed insertOne issue with PostgresDocumentManager that had no returned id

+++ src/tests/models/testModelsHelper.ts
@@ -5,6 +5,11 @@ const cleanupCollections = async () => {
     const colletionsToDrop = ['tests', 'testsWorker'];
 
     for (const collection of colletionsToDrop) {
+
+        if(!await App.container('db').schema().tableExists(collection)) {
+            continue;
+        }
+        
         await App.container('db').documentManager().table(collection).truncate();
     }
 }
